This commit improves code quality by refactoring and enhancing readability
- Abstracted TaskList to separate Class
- Shortened method lengths by introducing inner methods

These changes aim to make the codebase more maintainable, readable, and easier to understand, facilitating future development and debugging.
